requirements.txt
aiogram==3.2.0
pillow
moviepy
sqlite3
